Allow restarting services without dependencies

+++ src/service_spawner/service.cc
@@ -424,7 +424,7 @@ void SvdService::configureSlot() {
 }
 
 
-void SvdService::startSlot() {
+void SvdService::startSlot(bool withoutDeps) {
     logDebug() << "Invoked start slot for service:" << name;
     uptime.start();
 
@@ -463,7 +463,7 @@ void SvdService::startSlot() {
         }
 
         /* configure all dependencies before continue */
-        if (not config->dependencies.isEmpty()) {
+        if (not config->dependencies.isEmpty() && not withoutDeps) {
             Q_FOREACH(auto dependency, config->dependencies) {
                 logInfo() << "Installing and configuring dependency:" << dependency;
                 auto depConf = new SvdServiceConfig(dependency);
@@ -480,7 +480,7 @@ void SvdService::startSlot() {
         }
 
         /* after successful installation of core app and configuring dependencies, we may proceed */
-        if (not config->dependencies.isEmpty()) {
+        if (not config->dependencies.isEmpty() && not withoutDeps) {
             QFile::remove(indicator);
             logInfo() << "Found additional igniter dependency(ies) for service:" << name << "list:" << config->dependencies;
 
@@ -552,6 +552,11 @@ void SvdService::startSlot() {
     /* invoke after start slot */
     logTrace() << "After process start execution:" << name;
     emit afterStartSlot();
+}
+
+
+void SvdService::startSlot() {
+    startSlot(false);
 }
 
 
@@ -633,18 +638,20 @@ void SvdService::afterStartSlot() {
 }
 
 
-void SvdService::stopSlot() {
+void SvdService::stopSlot(bool withoutDeps) {
     logDebug() << "Invoked stop slot for service:" << name;
     auto config = new SvdServiceConfig(name);
     QString indicator = config->prefixDir() + DEFAULT_SERVICE_RUNNING_FILE;
     stopSitters();
 
     /* stop dependency services */
-    Q_FOREACH(SvdService *depService, this->dependencyServices) {
-        if (depService) {
-            logDebug() << "Invoking stop slot of service dependency:" << depService->name << "with uptime:" << toHMS(depService->getUptime());
-            depService->stopSlot();
-            depService->exit();
+    if (not withoutDeps) {
+        Q_FOREACH(SvdService *depService, this->dependencyServices) {
+            if (depService) {
+                logDebug() << "Invoking stop slot of service dependency:" << depService->name << "with uptime:" << toHMS(depService->getUptime());
+                depService->stopSlot();
+                depService->exit();
+            }
         }
     }
 
@@ -698,6 +705,10 @@ void SvdService::stopSlot() {
 }
 
 
+void SvdService::stopSlot() {
+    stopSlot(false);
+}
+
 void SvdService::afterStopSlot() {
     logDebug() << "Invoked after stop slot for service:" << name;
     logTrace() << "Loading service igniter" << name;
@@ -724,17 +735,22 @@ void SvdService::afterStopSlot() {
 }
 
 
-void SvdService::restartSlot() {
+void SvdService::restartSlot(bool withoutDeps) {
     if (QFile::exists(getHomeDir() + DEFAULT_SS_SHUTDOWN_HOOK_FILE)) {
         logWarn() << "Ignoring restart slot of service:" << name << "cause shutdown hook was called.";
     } else {
         logDebug() << "Invoked restart slot for service:" << name;
         // usleep(DEFAULT_SERVICE_PAUSE_INTERVAL);
         logWarn() << "Restarting service:" << name;
-        emit stopSlot();
-        emit startSlot();
+        emit stopSlot(withoutDeps);
+        emit startSlot(withoutDeps);
         logInfo() << "Service restarted:" << name;
     }
+}
+
+
+void SvdService::restartSlot() {
+    restartSlot(false);
 }
